requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
requests==2.22.0
Django==2.2.4
django-allauth==0.38.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.3
django-sslserver==0.20
django-storages==1.7.1
djangorestframework==3.10.2
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
idna==2.8
mpyq==0.2.5
oauthlib==3.0.1
psycopg2==2.8.4
pyasn1==0.4.5
pyasn1-modules==0.2.4
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
sentry-sdk==0.11.2
zephyrus_sc2_parser==0.2.22
google-cloud-pubsub==1.2.0
scipy==1.4.1
requests-futures==1.0.0
